+++ src/modules/media/configs/media.config.ts
