+++ src/modules/media/interfaces/media-rule.interface.ts
@@ -1,4 +1,4 @@
-import { RuleAt } from '../../core/rule/interfaces/rule.interface';
+import { RuleAt } from '../../../core/rule/interfaces/rule.interface';
 import { MediaBreakpoint } from './media.interface';
 
 export type MediaRuleDevice = 'mobile' | 'desktop';
