+++ src/modules/media/interfaces/media.interface.ts
